feat: export admin response types and add firefox handler
